requirements.txt
langchain
sentence_transformers
weaviate-client
pandas
python-env
